Coastal models

DIVACoast.jl also provides the higher-level data structures of Coastal Models, which provide a range of higher-level convenience functions to handle large ensembles of coastal risk assessment. The data structure of CoastalModel thereby combines hazard, exposure, vulnerabilty information for a given type of hazard. Several Coastal Models can further be combined into a CompositeCoastalModel. Currently, the only type of Coastal Model available is the CoastalFloodModel, which is further described below. Future versions of the library will also contain other types of Coastal Models such as, e.g., CoastalErosionModel and CoastalWetlandsModel.

Coastal Flood model

A CoastalFloodModel combines all information necessary for computing flood exposure and damage including sea-level hazard, attenuation model, exposure and vulnerability.

Main.DIVACoast.LocalCoastalImpactModel — Type

LocalCoastalImpactModel{DT<:Real, IDT, DATA} <: CoastalImpactUnit

A LocalCoastalImpactModel combines a surge model (distribution) and a coastal plain model (HypsometricProfile). It also holds the current protection level.

Main.DIVACoast.ComposedImpactModel — Type

ComposedImpactModel{ID_TYPE1,ID_TYPE2,DATA,CIU} <: CoastalImpactUnit where {CIU <: CoastalImpactUnit}

A ComposedImpactModel combines multiple LocalCoastalImpactModels into a combined datastructure.

A CoastalFloodModel is defined as

mutable struct CoastalFloodModel{DT<:Real,IDT,DATA} <: CoastalImpactUnit
  id::IDT
  surge_model::Distribution
  coastal_plain_model::HypsometricProfile{DT}
  protection_level::Real
  data::DATA
end