jdiva

provided by: Global Climate Forum

jdiva is a julia library for economic modelling of coastal sea-level rise impacts and adaptation. It provides a complete tool chain from geodatatypes to datatypes that allow different approaches of coasplain modelling to algorithms that compute flood impacts, erosion and wetland change.

GeoUtils

SparseGeoArray

lightweight data structure to store geodata

Main.jdiva.SparseGeoArray — Type

Base.getindex — Function

getindex(sga::SparseGeoArray, i::AbstractRange, j::AbstractRange, k::Union{Colon,AbstractRange,Integer})

Index a SparseGeoArray with AbstractRanges to get a cropped SparseGeoArray with the correct AffineMap set.

Examples

julia> sga[2:3,2:3]
2x2x1 Array{Float64, 3} with AffineMap([1.0 0.0; 0.0 1.0], [1.0, 1.0]) and undefined CRS

Main.jdiva.coords — Function

coords(sga::SparseGeoArray, p::SVector{2,<:Integer}, strategy::AbstractStrategy=Center())
coords(sga::SparseGeoArray, p::Tuple{<:Integer,<:Integer}, strategy::AbstractStrategy=Center())
coords(sga::SparseGeoArray, p::CartesianIndex{2}, strategy::AbstractStrategy=Center())

Retrieve coordinates of the cell index by p. See indices for the inverse function.

Main.jdiva.indices — Function

indices(sga::SparseGeoArray, p::SVector{2,<:Real})

Retrieve logical indices of the cell represented by coordinates p. See coords for the inverse function.

Geospatial analysis

tools to run geopspatial analysis

Main.jdiva.nh4 — Function

nh4(sga :: SparseGeoArray{DT, IT}, x :: Integer, y :: Integer) :: Array{Tuple{IT,IT}}
nh4(sga :: SparseGeoArray{DT, IT}, p :: Tuple{Integer,Integer}) :: Array{Tuple{IT,IT}}

Compute the 4-Neighbourhood of the grid cell x,y in the SparseGeoArray sga and return as Array of pairs. Takes into account the boundaries of the SparseGeoArray.

Examples

julia> nh4(sga, 1, 1)
2-element Vector{Tuple{Int32, Int32}}:
 (2, 1)
 (1, 2)

julia> nh4(sga, 2, 4)
4-element Vector{Tuple{Int32, Int32}}:
 (1, 4)
 (3, 4)
 (2, 3)
 (2, 5)

Main.jdiva.distance — Function

distance(lon1 :: R, lat1 :: R, lon2 :: R, lat2 :: R) :: R where {R <: Real}
distance(p1 :: SVector{2,R}, p2 :: SVector{2,R}) :: R where {R <: Real}
distance(p1 :: AbstractVector{R}, p2 :: AbstractVector{R}) :: R where {R <: Real} 
distance(p1 :: Tuple{R,R}, p2 :: Tuple{R,R}) where {R <: Real} :: R where {R <: Real}

Compute the distance (in km) between two points given by lon1,lat1 and lon2,lat2 resp. p1 and p2. Uses the Haversine formula.

Examples

julia> ...

Main.jdiva.go_direction — Function

go_direction(lon :: R, lat :: R, distance :: Real, direction :: AbstractDirection) :: Tuple{R,R} where {R <: Real}

Compute the geographical coordinates of the point reached if we go distance km from (lon,lat) in direction. Takes into account circularity, but does not cross poles. direction can be East(), North(), West(), South()

Examples

julia> go_direction(13.2240, 52.3057, 10, East())
(13.370916039175427, 52.3057)
julia> go_direction(13.2240, 52.3057, 10000, North())
(13.224, 90.0)
julia> go_direction(19.0045,0.0,40075,West())
(19.004500000000007, 0.0)

Main.jdiva.bounding_boxes — Function

bounding_boxes(sga :: SparseGeoArray{DT, IT}, lon_east :: Real, lon_west :: Real, lat_south :: Real, lat_north :: Real) where {DT <: Real, IT <: Integer}

Compute the bounding box(es) for the sparse geoarray sga and an area from loneast to lonwest and latsouth and latnorth.

Examples

julia> ...

Main.jdiva.bbox! — Function

Set geotransform of SparseGeoArray by specifying a bounding box. Note that this only can result in a non-rotated or skewed GeoArray.

Main.jdiva.is_rotated — Function

Check wether the AffineMap of a GeoArray contains rotations.

Geodata handling

tools to handle geodata

Main.jdiva.epsg! — Function

Set CRS on SparseGeoArray by epsgcode

Main.jdiva.proj2wkt — Function

Get the WKT of an Proj string

Main.jdiva.epsg2wkt — Function

Get the WKT of an Integer EPSG code

Main.jdiva.str2wkt — Function

Parse CRS string into WKT.

netCDF

utitlities to work with netCDF-files

Main.jdiva.load_hsps_nc — Function

load_hsps_nc(::Type{IT}, ::Type{DT}, filename::String)::Dict{IT,HypsometricProfile{DT}} where {IT<:Integer,DT<:Real}

Load a netcdf file into hypsometric profiles. IT is the indextype (an integer type which is used to adress a specific profile), DT is the datatype (a floating point type which is used to store the data internally). Hypsometric profiles are not compressed.

Examples

julia> test = load_hsps_nc(Int32, Float32, "test.nc")
Dict{Int32, HypsometricProfile{Float32}} with 2716 entries:
  2108 => HypsometricProfile{Float32}(1.0, Float32[-5.0, -4.9, -4.8, -4.7, -4.6, -4.5, -4.4, -4.3, -4.2, -4.1  …  19.1, 19.2, 19.3, 19.4, 19.5, 19.6, 19.7, 19.8, 19.9, 20.0], Float32[0.0, 0.0, 0.0, 0.0, …