Flood damage assessment

Flood propagation model

Currently DIVACoast.jl supports the bathtub model and the attenuated bathtub model. Attenuation refers to the reduction of water levels while floods propagate inland across the landscape. The magnitude of attenuation is a function of land cover such as vegetation, buildings and infrastructure which slow down and hence reduce the extent of flooding.

Main.DIVACoast.expected_damage_bathtub_standard_ddf — Function

expected_damage_bathtub_standard_ddf(LocalCoastalModel::LocalCoastalModel{DT}, hdd_area::DT, hdds_other::Array{DT})

This function calculates the annual expected damage for one local coastal model (Hypsometric Profile and Extreme surge distribution) by integrating the product of damages and the pdf (probability disctribution function) of the surge model over all possible extreme values. The output are annual expected damage (as a pair) for area (one number) and for all other exposure dimensions (an array of numbers). The standard depth damage function (dd = 1/(1+hdd)) is used to estimate flood damages, the hdd parameters for ares (one number) and for all other exposure dimensions (an array of numbers) are input parameters

Main.DIVACoast.expected_damage_bathtub — Function

expected_damage_bathtub(LocalCoastalModel::LocalCoastalModel{DT}, ddf_area::Function, ddf_other::Array{Function})

This function calculates the (annual) expected damage for one local coastal model (Hypsometric Profile and Extreme surge distribution) by integrating the product of damages and the pdf (probability disctribution function) of the surge model. The output are annual expected damage for (one number) and for all other exposure dimensions (an array of numbers). The depth damage functions for ares (one function) and for all other exposure dimensions (an array of functions) are input parameters.

Main.DIVACoast.damage_bathtub_standard_ddf — Function

Main.DIVACoast.damage_bathtub — Function

damage_bathtub(hspf::HypsometricProfile{DT}, wl::DT, ddf_area::Function, ddfs_other::Vector{Function}) where {DT<:Real}
damage_bathtub(hspf::HypsometricProfile{DT}, wl::DT, ddf_area::Function, ddfs_other::Vector{Any}) where {DT<:Real}
damage_bathtub(hspf::HypsometricProfile{DT}, wl::T, ddf::Function, s::Symbol) where {DT<:Real,T<:Real}

Calculates the damage using the bathtub model for a given water level wl based on a provided damage function ddfarea and additional damage functions ddfsother.

Arguments

hspf::HypsometricProfile{DT}: The hypsometric profile.

wl::DT: The water level for which the damage is calculated.

ddf_area::Function: The damage function for the area.

ddfs_other::Vector{Function}: Additional damage functions for other exposures.

Example

  damage = damage_bathtub(hspf, 1.5, (x) -> x^2, [(x) -> x, (x) -> x^3])