Drivers

DIVACoast.jl provides convenient data readers for external drivers such as sea-level rise and socio-economic development. These readers provide values for any future point in time by interpolating piecewise linearly between time steps and extrapolating linearly from the last available time step. All readers also provide growth rates between two points in time. Growth rates can be returned in three different ways as AnnualGrowthPercentage, AnnualGrowth, GrowthFactor. The readers can be used with the get_value() and get_value_from_cell() functions to retrieve the relevant data. The reader takes a NetCDF file as input, which must include the following dimensions:

Variable: The specific variable you want to access (e.g., Sea Level Rise in meters).

Longitude and Latitude: Dimensions specifying the longitude and latitude for each grid cell in the NetCDF file.

Time: Time dimension, e.g., 5-year increments.

Quantiles: Quantiles associated with the variable.

Main.DIVACoast.SSPScenarioReader — Type

Wraps SSP Dataset (CSV file).

Main.DIVACoast.SLRScenarioReader — Type

Creates a SLR-Scenario reader around a dataset (NetCDF) using dataset specific variable names for variable (e.g., "SeaLevelRise"), latitude, longitude, time, and quantile After the Wrapper structure was initialized dataset specific functions can be used (e.g, getslrvalue)

Main.DIVACoast.get_slr_value — Function

Gets the Sea Level Rise value at a specific location (lon, lat) in a specific quantile at a specific time.

Main.DIVACoast.get_slr_value_from_grid_cell — Function

Gets the Sea Level Rise value at a specific cell (indexlon, indexlat) in a specific quantile (given by index_qtl) at a specific time. Faster than the previous