Sea-Level hazard

Currently sea-level hazards in DIVACoast.jl are represented trough extreme still water level distributions. These distributions are implemented as Distributions of the Julia Package Distributions.jl. As ESL are often provided as non-parametric (i.e. empirical) distributions, i.e. point-wise as a list of water levels and associated return periods, DIVACoast.jl provides a couple of functions that convert these non-parametric distributions to a chosen parametric extreme value distribution.

Main.DIVACoast.estimate_gumbel_distribution — Function

This function tries to fit a gumbel distribution to given data. wldata is the actual data (e.g. water level). cdfdata are the empirical cdf values for the data in x (i.e. values between 0 and 1 - to be interpreted as quantiles). The funtion returns a GeneralizedExtremeValue (GEV) with the third (shape, ξ) parameter being zero. If the cdf fit fails for any reason, the standard gumbel distribution (μ=mean(data), σ=var(data), ξ=0.0) is returned

Main.DIVACoast.estimate_frechet_distribution — Function

This function fits a Frechet Distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV).

This function tries to fit a Frechet distribution to given data. x is the actual data (e.g. water level). y are the empirical cdf values for the data in x (i.e. values between 0 and 1 - to be interpreted as quantiles). The funtion returns a GeneralizedExtremeValue (GEV) with the third (shape, ξ) parameter being bigger than zero. If the cdf fit fails for any reason, a standard Frechet distribution (μ=mean(data), σ=var(data), ξ=0.5) is returned

Main.DIVACoast.estimate_gev_distribution — Function

This function fits an extreme value distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV) and uses the best fit out of the Gumbel, Frechet and Weibull model based on the summed squared residuals.

Main.DIVACoast.estimate_weibull_distribution — Function

This function fits a Weibull Distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV).

This function tries to fit a Weibull distribution to given data. 
    x is the actual data (e.g. water level).
    y are the cempirical df values for the data in x (i.e. values between 0 and 1 - to be interpreted as quantiles). 
The funtion returns a GeneralizedExtremeValue (GEV) with the third (shape, ξ) parameter being smaller than zero. If the cdf fit fails for any reason, 
a standard Weibull distribution (μ=mean(data), σ=var(data), ξ=-0.5) is returned

Main.DIVACoast.estimate_gpd_negative_distribution — Function

This function fits a Generalized Pareto Distribution with negative shape to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedPareto (GPD).

Main.DIVACoast.estimate_gpd_positive_distribution — Function

This function fits a gpd_positive Distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedExtremeValue (GEV).

Main.DIVACoast.estimate_gp_distribution — Function

This function fits an extreme value distribution to the inserted data. y should be the return period and x the corresponding water level height. The funtion returns a GeneralizedPareto (GPD) and uses the best fit out of the exponential, positive GPD and negative GPD model based on the summed squared residuals.

Main.DIVACoast.estimate_exponential_distribution — Function

This function tries to fit an exponential distribution to given data. x is the actual data (i.e. water level). y are the cdf values for the data in x (i.e. values between 0 and 1). The funtion returns a GeneralizedPareto (GPD) with the third shape parameter being zero. If the cdf fit fails for any reason, the standard exponential distribution (μ=0) is returned

Main.DIVACoast.plot_comparison_extreme_distributions — Function

plot_comparison_extreme_distributions(x_data::Array{T}, y_data::Array{T})

This function creates a plot that compares six different extreme distribution fits to underlying data. The input is an xdata and ydata array, xdata is the absolute height of extremes and ydata is the annual probability to be below this value. The function fits three different GEV and three different GPD functions to the data and plots all of these. It also highlights the best fit out of these six functions and highlights this in the plot.